Avoid potential undefined behaviour
This code worked correctly in many compilers (with quiet assumption,
that Chip and Channel are standard layout types), but it has undefined
behaviour according to the C standard, since it involves a dereference
of a null pointer.

Use C++11 static_assert to assure, that no UB is triggered (offsetof on
non-std-layout types does not trigger UB only starting with C++17).

Use offsetof constexpr instead of dereferencing nullptr in runtime to
compute field offset.